docs(examples): move geometry data to yml file, inline utilities
